Migrate PostFrontMatter to server component
