+++ src/blog/[slug]/share-buttons/social-share-buttons.tsx
@@ -1,0 +1,40 @@
+import { LinkedinIcon, TwitterIcon } from "../../../shared/icons";
+import { PostMeta } from "../../../shared/types";
+import { getPostUrl } from "../../../shared/urls";
+import { ShareButton } from "./share-button";
+
+export function ShareOnTwitterButton({ post }: ShareOnTwitterButtonProps) {
+  return (
+    <ShareButton
+      href={`https://twitter.com/intent/tweet?url=${encodeURIComponent(
+        getPostUrl(post.slug)
+      )}`}
+      rel="noreferrer"
+      target="_blank"
+      title="Share on Twitter"
+      icon={TwitterIcon}
+    />
+  );
+}
+
+export type ShareOnTwitterButtonProps = {
+  post: PostMeta;
+};
+
+export function ShareOnLinkedinButton({ post }: ShareOnLinkedinButtonProps) {
+  return (
+    <ShareButton
+      href={`https://www.linkedin.com/sharing/share-offsite/?url=${encodeURIComponent(
+        getPostUrl(post.slug)
+      )}`}
+      rel="noreferrer"
+      target="_blank"
+      title="Share on LinkedIn"
+      icon={LinkedinIcon}
+    />
+  );
+}
+
+export type ShareOnLinkedinButtonProps = {
+  post: PostMeta;
+};
